Save language to localStorage and load product information in product-form
