+++ src/app/components/features/product-form/auction-form/auction-form.ts
