+++ src/app/components/features/product-form/product-general-form/product-general-form.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ProductGeneralForm } from './product-general-form';
+
+describe('ProductGeneralForm', () => {
+  let component: ProductGeneralForm;
+  let fixture: ComponentFixture<ProductGeneralForm>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ProductGeneralForm]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(ProductGeneralForm);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
