+++ src/app/components/features/products-list/product-list-view/product-item/product-item.html
@@ -22,7 +22,7 @@
             </div>
             <div class="flex flex-col gap-2">
               <p class="text-base md:text-xl">{{ 'product_details.localization' | translate }}: {{ product()?.location?.country }}, {{product()?.location?.city}}, {{product()?.location?.state}}</p>
-              <p class="text-base md:text-xl">{{ 'product_details.date_of_issue' | translate }}: {{ product()?.dateCreated }}</p>
+              <p class="text-base md:text-xl">{{ 'product_details.date_of_issue' | translate }}: {{ product()?.dateCreated | date:'yyyy-MM-dd' }}</p>
             </div>
           </div>
           <div class="flex-none flex flex-col items-start md:items-end md:pr-8">
