+++ src/app/components/features/user-products/user-product-item/user-product-item.html
@@ -1,5 +1,5 @@
-<a routerLink="/product-form" [queryParams]="{ id: product()?.id, type: product()?.listingType }" class="w-full max-w-full">
-  <div class="flex flex-col w-full border-b-1 p-4">
+<a routerLink="/product-form" [queryParams]="{ id: product()?.id, listingType: product()?.listingType }" class="w-full max-w-full">
+  <div class="flex bg-background flex-col w-full border-b-1 hover:border-b-transparent p-4 hover:rounded-xl hover:shadow-xl hover:z-10 hover:scale-[1.02] duration-150 ">
     @if (isMd()){
       <p class="text-xl md:text-4xl pb-4">{{ product()?.name }}</p>
     }
@@ -16,18 +16,24 @@
             <div class="flex flex-row flex-wrap gap-3">
               @for (category of product()?.categories; track category.id){
                 <div class="border-1 px-4 py-2 rounded-2xl">
-                  <p class="text-base md:text-xl">{{category.name}}</p>
+                  <p class="text-base md:text-xl">{{ `categories.${category.name}` | translate }}</p>
                 </div>
               }
             </div>
             <div class="flex flex-col gap-2">
-              <p class="text-base md:text-xl">Localization: {{ product()?.location?.country }}, {{product()?.location?.city}}, {{product()?.location?.state}}</p>
-              <p class="text-base md:text-xl">Date of issue: {{ product()?.dateCreated }}</p>
+              <p class="text-base md:text-xl">{{ 'product_details.localization' | translate }}: {{ product()?.location?.country }}, {{product()?.location?.city}}, {{product()?.location?.state}}</p>
+              <p class="text-base md:text-xl">{{ 'product_details.date_of_issue' | translate }}: {{ product()?.dateCreated | date:'yyyy-MM-dd' }}</p>
             </div>
           </div>
           <div class="flex-none flex flex-col items-start md:items-end md:pr-8">
-            <p class="text-xl md:text-3xl">{{ product()?.price }}$</p>
-            <p class="text-base md:text-lg">{{ product()?.isNegotiable ? "negotiable" : "" }}</p>
+            @if(product()?.listingType === LISTING_TYPES.PRODUCT){
+              <p class="text-xl md:text-3xl">{{ 'product_details.price' | translate }}: {{ product()?.price }} $</p>
+              <p class="text-base md:text-lg">{{ product()?.isNegotiable ? ('product_details.negotiable' | translate) : "" }}</p>
+            }
+            @else if(product()?.listingType === LISTING_TYPES.AUCTION){
+              <p class="text-xl md:text-3xl">{{ 'product_details.current_bid' | translate }}: {{ product()?.currentBid ?? product()?.startingBid}}$</p>
+              <p class="text-base md:text-lg">{{ 'product_details.starting_bid' | translate }}: {{ product()?.startingBid }}</p>
+            }
           </div>
         </div>
       </div>
