+++ src/app/components/features/user-products/user-product-item/user-product-item.ts
@@ -1,11 +1,16 @@
 import {Component, HostBinding, input} from '@angular/core';
 import {ListingItemDTO} from '../../../../core/models/LitstingItemDTO';
 import {RouterLink} from '@angular/router';
+import {TranslatePipe} from '@ngx-translate/core';
+import {LISTING_TYPES} from '../../../../core/constants/constants';
+import {DatePipe} from '@angular/common';
 
 @Component({
   selector: 'app-user-product-item',
   imports: [
-    RouterLink
+    RouterLink,
+    TranslatePipe,
+    DatePipe
   ],
   templateUrl: './user-product-item.html',
   styleUrl: './user-product-item.css'
@@ -15,4 +20,5 @@ export class UserProductItem {
 
   readonly isMd = input.required<boolean>();
   readonly product = input<ListingItemDTO>();
+  protected readonly LISTING_TYPES = LISTING_TYPES;
 }
