+++ src/app/components/features/user-products/user-products.html
@@ -1,7 +1,7 @@
 <div class="flex m-auto flex-col max-w-7xl w-full py-4 px-8 gap-4">
-  <h1>Your products and auctions</h1>
+  <h1>{{ 'user_products.header' | translate }}</h1>
   <div class="w-full flex justify-between py-4 gap-2 flex-wrap">
-    <app-button label="Create" (clickButton)="openDialog()"/>
+    <app-button [label]="'create' | translate" (clickButton)="openDialog()"/>
     @if(!loading() && totalPages() !== 0){
       <app-paginator [isBlocked]="isBlocked()" [pageNumber]="filter().pageNumber" [totalPages]="totalPages()" (pageChange)="usePaginator($event)"/>
     }
@@ -11,12 +11,12 @@
         <app-loader/>
       </div>
     } @else if (totalPages() === 0){
-      <div class="flex justify-center py-16  w-full">
+      <div class="flex justify-center py-16 w-full">
         <h2>Can't find any article. Create your first product!</h2>
       </div>
     }
     @else {
-      <div class="flex flex-col gap-4 w-full">
+      <div class="flex flex-col w-full">
         @let data = listings();
 
         @if(data && data.items.length){
@@ -35,11 +35,11 @@
   <app-dialog (closed)="closeDialog()">
     <div class="flex flex-col justify-center h-full w-full gap-4">
       <div class="px-16 flex flex-row gap-4">
-        <app-button label="Product" (clickButton)="goToProductForm(LISTING_TYPES.PRODUCT)"/>
-        <app-button label="Auction" severity="secondary" (clickButton)="goToProductForm(LISTING_TYPES.AUCTION)"/>
+        <app-button [label]="'listing_types.product' | translate" (clickButton)="goToProductForm(LISTING_TYPES.PRODUCT)"/>
+        <app-button [label]="'listing_types.auction' | translate" severity="secondary" (clickButton)="goToProductForm(LISTING_TYPES.AUCTION)"/>
       </div>
       <div class="flex flex-row">
-        <app-button label="Cancel" severity="accent" variant="outlined" (clickButton)="closeDialog()"/>
+        <app-button [label]="'cancel' | translate" severity="accent" variant="outlined" (clickButton)="closeDialog()"/>
       </div>
     </div>
   </app-dialog>
